# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NudgeLab API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/trainings": {
      "get": {
        "tags": [
          "studio"
        ],
        "summary": "List Trainings",
        "operationId": "list_trainings_api_v1_trainings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TrainingListItem"
                  },
                  "type": "array",
                  "title": "Response List Trainings Api V1 Trainings Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "studio"
        ],
        "summary": "Create Training",
        "operationId": "create_training_api_v1_trainings_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrainingCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrainingDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TrainingCreate": {
        "properties": {
          "training_id": {
            "type": "string",
            "pattern": "^[a-z][a-z0-9_]{2,49}$",
            "title": "Training Id"
          },
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "completion_type": {
            "type": "string",
            "enum": [
              "quiz",
              "walkthrough",
              "acknowledgment"
            ],
            "title": "Completion Type",
            "default": "quiz"
          },
          "uses_location": {
            "type": "boolean",
            "title": "Uses Location",
            "default": false
          },
          "trainer_name": {
            "type": "string",
            "maxLength": 40,
            "minLength": 1,
            "title": "Trainer Name",
            "default": "Anne"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 30
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "completion_key": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^[A-Za-z0-9_-]{1,64}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "training_id",
          "title"
        ],
        "title": "TrainingCreate"
      },
      "TrainingDetail": {
        "properties": {
          "training_id": {
            "type": "string",
            "title": "Training Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "completion_type": {
            "type": "string",
            "enum": [
              "quiz",
              "walkthrough",
              "acknowledgment"
            ],
            "title": "Completion Type"
          },
          "uses_location": {
            "type": "boolean",
            "title": "Uses Location"
          },
          "completion_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Key"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "app_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Title"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required"
          },
          "app_status": {
            "type": "string",
            "title": "App Status"
          },
          "wanaka_trainer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wanaka Trainer Id"
          },
          "active_version_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Active Version Id"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/VersionSummary"
            },
            "type": "array",
            "title": "Versions"
          }
        },
        "type": "object",
        "required": [
          "training_id",
          "title",
          "status",
          "completion_type",
          "uses_location",
          "completion_key",
          "profile_id",
          "app_title",
          "category",
          "description",
          "tags",
          "is_required",
          "app_status",
          "wanaka_trainer_id",
          "active_version_id",
          "versions"
        ],
        "title": "TrainingDetail"
      },
...
      "TrainingListItem": {
        "properties": {
          "training_id": {
            "type": "string",
            "title": "Training Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "completion_type": {
            "type": "string",
            "enum": [
              "quiz",
              "walkthrough",
              "acknowledgment"
            ],
            "title": "Completion Type"
          },
          "uses_location": {
            "type": "boolean",
            "title": "Uses Location"
          },
          "completion_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completion Key"
          },
          "profile_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Profile Id"
          },
          "app_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "App Title"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "is_required": {
            "type": "boolean",
            "title": "Is Required"
          },
          "app_status": {
            "type": "string",
            "title": "App Status"
          },
          "wanaka_trainer_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wanaka Trainer Id"
          },
          "active_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VersionRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "drafts": {
            "type": "integer",
            "title": "Drafts"
          },
          "in_review": {
            "type": "integer",
            "title": "In Review"
          },
          "last_edited_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Edited At"
          }
        },
        "type": "object",
        "required": [
          "training_id",
          "title",
          "status",
          "completion_type",
          "uses_location",
          "completion_key",
          "profile_id",
          "app_title",
          "category",
          "description",
          "tags",
          "is_required",
          "app_status",
          "wanaka_trainer_id",
          "active_version",
          "drafts",
          "in_review",
          "last_edited_at"
        ],
        "title": "TrainingListItem"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "VersionRef": {
        "properties": {
          "version_id": {
            "type": "integer",
            "title": "Version Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          }
        },
        "type": "object",
        "required": [
          "version_id",
          "label"
        ],
        "title": "VersionRef"
      },
...
      "VersionSummary": {
        "properties": {
          "version_id": {
            "type": "integer",
            "title": "Version Id"
          },
          "training_id": {
            "type": "string",
            "title": "Training Id"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "status": {
            "type": "string",
            "enum": [
              "draft",
              "in_review",
              "published",
              "retired"
            ],
            "title": "Status"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "revision": {
            "type": "integer",
            "title": "Revision"
          },
          "has_content": {
            "type": "boolean",
            "title": "Has Content"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "updated_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated By"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          }
        },
        "type": "object",
        "required": [
          "version_id",
          "training_id",
          "label",
          "status",
          "is_active",
          "revision",
          "has_content",
          "notes",
          "created_at",
          "created_by",
          "updated_at",
          "updated_by",
          "published_at"
        ],
        "title": "VersionSummary"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}